moving Settings and Version to package object
